Killzone + Levels
Added the killzone in which the ball resets at the bottom of the screen, taking the players health away which is capped at 3. As well as adding the Level restart after losing on life which resets ball and paddle position and a game over which makes the game restart from level 1.  Added a second level for testing sake, to have a simple level switcher after clearing each board